correcting feature grid unit test file location
